+++ CHANGES.md
@@ -12,10 +12,10 @@ To be released.
  -  (Libplanet.Explorer) Changed `BencodexValueType` to inherit
     `ObjectGraphType<IValue>` instead of `StringGraphType`.  Instead of
     simply being a hexadecimal representation of `byte[]` encoded `IValue`,
-    now one can choose its representation format.  [[#3551]]
+    now one can choose its representation format.  [[#3560]]
 
-[#3559]: https://github.com/planetarium/libplanet/pull/3549
-[#3551]: https://github.com/planetarium/libplanet/pull/3551
+[#3559]: https://github.com/planetarium/libplanet/pull/3559
+[#3560]: https://github.com/planetarium/libplanet/pull/3560
 
 
 Version 3.9.2