+++ CHANGES.md
@@ -128,7 +128,12 @@ To be released.
 Version 3.9.2
 -------------
 
-To be released.
+Released on December 15, 2023.
+
+ -  (Libplanet.RocksDBStore) Added `DBOptions` type parameter to
+    `RocksDBKeyValueStore()`.  [[#3553]]
+
+[#3553]: https://github.com/planetarium/libplanet/pull/3553
 
 
 Version 3.9.1